WebUI: Policy-based generation of RucioUI pages #245
