+++ lib/rucio/web/ui/templates/atlas_index.html
@@ -7,9 +7,12 @@
  http://www.apache.org/licenses/LICENSE-2.0
 
  Authors:
- - Mario Lassnig, <mario.lassnig@cern.ch>, 2014
  - Thomas Beermann, <thomas.beermann@cern.ch>, 2017
 -->
 
 $var title: Index
-$var script: <script src="/static/index.js"></script>
+$var script: <script src="/static/atlas_index.js"></script>
+
+<center>
+  <div id="gridview" style="width: 50%; height: 50%;">[loading]</div>
+</center>
