+++ lib/rucio/web/ui/templates/base.html
@@ -1,4 +1,4 @@
-$def with (token, account, rucio_ui_version, content)
+$def with (token, account, rucio_ui_version, policy, content)
 <!--
  Copyright European Organization for Nuclear Research (CERN)
 
@@ -76,12 +76,12 @@ $def with (token, account, rucio_ui_version, content)
                 <li><a href="/subscriptions">Subscriptions &amp; Rules</a></li>
                 <li><a href="/rse_locks">RSE Locks</a></li>
                 <li><a href="/rse_usage">Rucio Storage Elements (RSE)</a></li>
-                <li><a href="/account_usage">Group Account Usage</a></li>
+                $if policy == 'atlas': <li><a href="/account_usage">Group Account Usage</a></li>
                 <!--<li><a href="/api_usage">API Usage</a></li>-->
-                <li><a href="/rule_backlog_monitor">Rule Backlog</a></li>
+                $if policy == 'atlas': <li><a href="/rule_backlog_monitor">Rule Backlog</a></li>
                 <li><a href="/bad_replicas/summary">Bad file replicas</a></li>
-                <li><a href="/dbrelease_summary">DBRelease Overview</a></li>
-                <li><a href="/conditions_summary">ConditionsDB Overview</a></li>
+                $if policy == 'atlas': <li><a href="/dbrelease_summary">DBRelease Overview</a></li>
+                $if policy == 'atlas': <li><a href="/conditions_summary">ConditionsDB Overview</a></li>
                 <!--<li><a>TBD: Rucio Accounts</a></li>-->
               </ul>
             </li>
@@ -94,12 +94,13 @@ $def with (token, account, rucio_ui_version, content)
               </ul>
             </li>
 
-            <li class="has-dropdown">
-              <a>Reports</a>
-              <ul class="dropdown">
-                <li><a href="/dumps">Dumps</a></li>
-              </ul>
-            </li>
+            $if policy == 'atlas':
+                <li class="has-dropdown">
+                   <a>Reports</a>
+                   <ul class="dropdown">
+                     <li><a href="/dumps">Dumps</a></li>
+                   </ul>
+                </li>
 
             <li id="admin_dropdown" class="has-dropdown" style="display:none;">
               <a>Admin</a>
@@ -135,26 +136,27 @@ $def with (token, account, rucio_ui_version, content)
               </ul>
             </li>
 
-            <li class="has-dropdown">
-              <a>Other Monitoring</a>
-              <ul class="dropdown">
-                <li><a href="http://dashb-atlas-data.cern.ch/ddm2/">DDM Dashboard</a>
-                <li><a href="http://dashb-fts-transfers.cern.ch/">FTS Dashboard</a>
-                <li><a href="https://fts3-pilot.cern.ch:8449/fts3/ftsmon/#/">FTS Monitor CERN</a>
-                <li><a href="https://www-ftsmon-test.gridpp.rl.ac.uk:8449/fts3/ftsmon/#/">FTS Monitor RAL</a>
-                <li><a href="http://bigpanda.cern.ch/">PanDA Monitor</a>
-                <li><a href="http://adc-monitoring.cern.ch/">ADC Monitoring</a>
-              </ul>
-            </li>
+            $if policy == 'atlas':
+              <li class="has-dropdown">
+                <a>Other Monitoring</a>
+                <ul class="dropdown">
+                  <li><a href="http://dashb-atlas-data.cern.ch/ddm2/">DDM Dashboard</a>
+                  <li><a href="http://dashb-fts-transfers.cern.ch/">FTS Dashboard</a>
+                  <li><a href="https://fts3-pilot.cern.ch:8449/fts3/ftsmon/#/">FTS Monitor CERN</a>
+                  <li><a href="https://www-ftsmon-test.gridpp.rl.ac.uk:8449/fts3/ftsmon/#/">FTS Monitor RAL</a>
+                  <li><a href="http://bigpanda.cern.ch/">PanDA Monitor</a>
+                  <li><a href="http://adc-monitoring.cern.ch/">ADC Monitoring</a>
+                </ul>
+              </li>
 
-            <li class="has-dropdown">
-              <a><span class="help" style="color:#b2d546">Help</span></a>
-              <ul class="dropdown">
-                <li><a href="mailto:atlas-adc-ddm-support@cern.ch">ADC DDM Support Mailing List</a>
-                <li><a href="mailto:rucio-dev@cern.ch">Rucio Development Mailing List</a>
-                <li><a href="http://rucio.cern.ch/">Rucio Web Site</a>
-              </ul>
-            </li>
+              <li class="has-dropdown">
+                <a><span class="help" style="color:#b2d546">Help</span></a>
+                <ul class="dropdown">
+                  <li><a href="mailto:atlas-adc-ddm-support@cern.ch">ADC DDM Support Mailing List</a>
+                  <li><a href="mailto:rucio-dev@cern.ch">Rucio Development Mailing List</a>
+                  <li><a href="http://rucio.cern.ch/">Rucio Web Site</a>
+                </ul>
+              </li>
           </ul>
         </section>
       </nav>
